Update to use verifier config
