+++ crates/sui-open-rpc/spec/openrpc.json
@@ -1353,6 +1353,7 @@
                 "accept_zklogin_in_multisig": false,
                 "advance_epoch_start_time_in_safe_mode": true,
                 "advance_to_highest_supported_protocol_version": false,
+                "allow_receiving_object_id": false,
                 "ban_entry_init": false,
                 "commit_root_state_digest": false,
                 "consensus_order_end_of_epoch_last": true,
@@ -1362,6 +1363,7 @@
                 "enable_effects_v2": false,
                 "enable_jwk_consensus_updates": false,
                 "end_of_epoch_transaction_supported": false,
+                "hardened_otw_check": false,
                 "include_consensus_digest_in_prologue": false,
                 "loaded_child_object_format": false,
                 "loaded_child_object_format_type": false,
