+++ external-crates/move/move-execution/v0/move-bytecode-verifier/bytecode-verifier-tests/src/unit_tests/mod.rs
@@ -52,5 +52,6 @@ pub(crate) fn production_config() -> VerifierConfig {
         max_per_mod_meter_units: Some(1000 * 8000),
         max_constant_vector_len: Some(DEFAULT_MAX_CONSTANT_VECTOR_LEN),
         max_idenfitier_len: Some(DEFAULT_MAX_IDENTIFIER_LENGTH),
+        allow_receiving_object_id: false,
     }
 }
