+++ sui-execution/latest/sui-adapter/src/programmable_transactions/execution.rs
@@ -58,6 +58,7 @@ mod checked {
         execution_status::{CommandArgumentError, PackageUpgradeError},
     };
     use sui_verifier::{
+        default_verifier_config,
         private_generics::{EVENT_MODULE, PRIVATE_TRANSFER_FUNCTIONS, TRANSFER_MODULE},
         INIT_FN_NAME,
     };
@@ -840,7 +841,7 @@ mod checked {
             sui_verifier::verifier::sui_verify_module_unmetered(
                 module,
                 &BTreeMap::new(),
-                context.protocol_config.version.as_u64(),
+                &default_verifier_config(context.protocol_config, false),
             )?;
         }
