+++ sui-execution/latest/sui-verifier/src/verifier.rs
@@ -4,6 +4,7 @@
 //! This module contains the public APIs supported by the bytecode verifier.
 
 use move_binary_format::file_format::CompiledModule;
+use move_vm_config::verifier::VerifierConfig;
 use sui_types::{error::ExecutionError, move_package::FnInfoMap};
 
 use crate::{
@@ -18,12 +19,12 @@ pub fn sui_verify_module_metered(
     module: &CompiledModule,
     fn_info_map: &FnInfoMap,
     meter: &mut impl Meter,
-    protocol_version: u64,
+    verifier_config: &VerifierConfig,
 ) -> Result<(), ExecutionError> {
     struct_with_key_verifier::verify_module(module)?;
     global_storage_access_verifier::verify_module(module)?;
     id_leak_verifier::verify_module(module, meter)?;
-    private_generics::verify_module(module, protocol_version)?;
+    private_generics::verify_module(module, verifier_config)?;
     entry_points_verifier::verify_module(module, fn_info_map)?;
     one_time_witness_verifier::verify_module(module, fn_info_map)
 }
@@ -35,10 +36,10 @@ pub fn sui_verify_module_metered_check_timeout_only(
     module: &CompiledModule,
     fn_info_map: &FnInfoMap,
     meter: &mut impl Meter,
-    protocol_version: u64,
+    verifier_config: &VerifierConfig,
 ) -> Result<(), ExecutionError> {
     // Checks if the error counts as a Sui verifier timeout
-    if let Err(error) = sui_verify_module_metered(module, fn_info_map, meter, protocol_version) {
+    if let Err(error) = sui_verify_module_metered(module, fn_info_map, meter, verifier_config) {
         if matches!(
             error.kind(),
             sui_types::execution_status::ExecutionFailureStatus::SuiMoveVerificationTimedout
@@ -53,9 +54,9 @@ pub fn sui_verify_module_metered_check_timeout_only(
 pub fn sui_verify_module_unmetered(
     module: &CompiledModule,
     fn_info_map: &FnInfoMap,
-    protocol_version: u64,
+    verifier_config: &VerifierConfig,
 ) -> Result<(), ExecutionError> {
-    sui_verify_module_metered(module, fn_info_map, &mut DummyMeter, protocol_version).map_err(
+    sui_verify_module_metered(module, fn_info_map, &mut DummyMeter, verifier_config).map_err(
         |err| {
             // We must never see timeout error in execution
             debug_assert!(
